Fix alpha for feet
Set ROI on grayImage to a narrow band above the pressure mat.
Find contours, and use these to draw ellipses that roughly matte in the subject's feet.
It's far from perfect, and we do get some weird artifacting / unwanted feet outside the mat area. Best to shrink the ROI to just above the pressure mat. But HEY, it's a start!